Numerous stability updates
- cmt/log is now thread-safe and prints the thread name
- it is now possible to cancel tasks mid-operation from futures.
- added a spinlock that does not yield
- moved yield,sleep, and sleep_until methods on thread interface to
  protected because they can only be called by current thread and that
  is ensured by boost::cmt::yield,sleep, and sleep_until.
- Fixed many bugs associated with exiting the program.

If you do a synchronous invoke from a non-cmt thread it will use the
blocking promise.  The blocking promise will throw if SIGINT is received
and therefore delete the stack-based promise and task objects which when
run will eventually crash.  It may be advisable to only use stack-based
objects with cmt based stacks.