Improved Cleanup of Threads
When one thread calls quit() on another it now
joins that thread.

Threads make sure that all cooperative stacks are unwound.

Added the ability to call thread::exec() from a coroutine when
you want that coroutine maintain the stack state until the program
exits.   This does reduce available space for other tasks to the
unused space of that task.