Actor Interface based on Boost.Reflect any_ptr
When an object 'lives' in one thread, all actions on that object need to
be re-posted to the proper thread.  This process can be tedious because
every method must check whether the current thread is the same as the
object's thread and if not, re-post the event in the proper thread.

boost::cmt::actor<T> abstracts this process by using
boost::reflect::any_ptr to adapt the interface to always post to the
proper thread.

- boost::cmt::log now prints diagnostic information when exceptions are
  thrown from args.